input: synaptics_i2c_rmi4: Calibrate on display on
If device is turned on with a finger on a panel during wireless
charging, touch could be in wrong baseline. Wrong baseline causes
ghost touch or nonreactive touch problem.
In case of this, we need an additional calibration after display
is turned on. If display is turned on by touch event, calibrate once
after the all touch events are released.

Bug: 35938075
Change-Id: Icd95d76772c60a151cc5c6401988fdf6d562bdab
Signed-off-by: jjhaaron.jeong <jjhaaron.jeong@lge.com>