binder: fix possible hang if debug interface is used
There is a timing window during creation of a new binder_proc
where the new proc is on the binder_procs list, but the
binder allocator mutex is not initialized. If someone
dumps out the binder stats via sysfs, it can hang trying
to acquire the uninitialized mutex. Fixed by initializing
binder_alloc before adding the new proc to binder_procs.

Test: tested manually
Bug: 32225111
Change-Id: I1e4b971d79f7b67c78642147c5e67ddec58207df
Signed-off-by: Todd Kjos <tkjos@google.com>