usb: typec: fix potential NULL pointer dereference
dwc3-msm driver registers uc and callbacks in the middle
of the probe, and can fail later in the probe. This leaves
uc to be an invalid pointer that can potentially cause NULL
pointer dereference. Move the uc callbacks registration at
the end of the probe so uc callbacks are valid and ready
when registered to fusb302 driver.

Also adds mutex protection around access to chip->uc and
chip->utc to prevent race.

Bug: 35795496
Change-Id: I643880226ec901e1ffb4e4075c560ac6e83f4f8a
Signed-off-by: Yueyao Zhu <yueyao@google.com>