KASAN: Panic on first invalid write
After the first instance of memory corruption occurs, the kernel
is no longer reliable and later reports of memory corruption could
just be side effects of the first one. Plus we want to panic to get
the real culprit of memory corruption in scenarios where serial
console isn't available, like on user's devices.

Bug: 36105810
Change-Id: I1ba75e4961afe653219cf81b4299f017c897d896
Signed-off-by: Siqi Lin <siqilin@google.com>