drivers/nfc/nq-nci: add qcom,postinit_cmd u8 array
Add an optional u8 to the device tree which specifies a
post-initialization command to issue over i2c in the format of

  qcom,nq-postinit-cmd = /bits/ 8
    <0x5 0x20 0x02 0x05 0x01 0xA0 0xF2 0x01 0x01>;

This can be used to initialize any specific NFC behavior in the kernel
before the NFC userland stack is available to do so.  This may be useful
in cases where the NFC controller is responsible for other attached
devices, like a SPI attached secure element, that should be made
accessible even without operating system NFC support.

Bug: 35105409
Test: works with pn81a driver and libese

Signed-off-by: Will Drewry <drewry@google.com>

Change-Id: I25651d97a75d270da92739d532e43e1f14203d79