smb-lib: remove otg_en variable as a cached hardware state
otg_en can be inconsistent with the actual hardware state in
some cases. For example, when _smblib_otg_enable() returns
-ETIMEDOUT, otg_en will not be set to true. The above case
can happen when the vbus source is switched from the external
booster to the internal booster, and PMIC booster will not
turn on until the VBUS voltage by the external booster is removed.
However, although -ETIMEDOUT is returned, when the voltage by the
external booster is removed, PMIC booster starts providing output,
i.e. OTG is enabled, but otg_en variable is still false.

This commit removes the otg_en variable as a cached hardware state,
and instead query directly for the hardware state.

Change-Id: I296ac54fbff5b192453e7e116c2db1e1c7cd4333
Signed-off-by: Yueyao Zhu <yueyao@google.com>