usb: ptn36241g: initialize CC1/CC2 pin state
When the device boots up without USB connected, chip->
ptn_active is initialized to false at the probe stage.
In psy_changed_handler(), as the cc_orientation is
none, and thus ptn_enable = false, and same as
chip->ptn_active. As a result, the pinctrl for
the chip is not properly initialized.

Bug:35025621
Change-Id: Iff55e0a642008dd280920651398e2c61804d5104
Signed-off-by: puma_hsu <puma_hsu@htc.com>
Signed-off-by: Yueyao Zhu <yueyao@google.com>