drv2624: convert to led-trig transient device
This patch converts the driver from an Android timed-output device to
led-trig transient device. One of the benefits is that it no longer has
to maintain it's own hrtimer. In addition to that, Android timed-output
is being removed after 4.9, this change would make it easier to port the
driver to a newer kernel version: https://patchwork.kernel.org/patch/8664831/

Bug: 36219950

Change-Id: I83c8b276d90b3cb0a64681f472654c28dac9b4b5
Signed-off-by: David Lin <dtwlin@google.com>