power: qpnp-charger: Modify the condition code for BATFET control.
BATFET must control NPM mode instead of LPM after WiFi on
otherwise the device might go off due to insufficient
current supply for WiFi.
BATFET control won't work if charging_disabled is true, which is
true until the boot completed. This causes the sudden power off
in case that WiFi is truned on at the meantime.

Bug: 35903443
Change-Id: I55ed64d79effe7555ecc711f6ef69ba013752459
Signed-off-by: kinam119.kim <kinam119.kim@lge.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>