smb-lib: continue overheat mitigation until cable unplugged
When the overheat mitigation is running, and the temperature
has dropped below the threshold, probe the cc status to see
if the cable is unplugged. Ends the mitigation only when
the cablue is unplugged.

The check is done by briefly enabling the Type-C block as
Type-C power role sink, wait for a delay to let the Type-C
block to settle and check the CC attachment status. Disable
the Type-C block at the end.

Bug: 35398639

Change-Id: Id9ed560682a77fc464ce40b94ef4ec768c5bfb35
Signed-off-by: Yueyao Zhu <yueyao@google.com>