smb-lib: hold wakelock when port overheat monitor is running
From the start of the overheat monitor (charging status change),
until the overheat monitor stops (i.e. none of the three conditions
are met, 1) the device is still being charged OR 2) overheat is
currently happending OR 3) an error has occurred), a wake lock is
held to prevent the system from suspend. Otherwise, the overheat
monitor work may not be run in time to perform mitigation.

TEST:
- charge the device (workstation USB)
- - monitor start running (smblib log: "Port overheat work running")
- echo 800 > /sys/class/power_supply/usb/port_temp
- - heat mitigation starts
- unplug cable
- - monitor still runs, system does not suspend
- echo 250 > /sys/class/power_supply/usb/port_temp
- - heat mitigation ends (log: "Port overheat mitigation ends.")
- - system goes to suspend

Bug: 35398639
Change-Id: I3c7eafc392d26d275245c9022fc46ff6b4f084a4
Signed-off-by: Yueyao Zhu <yueyao@google.com>