2017-05-partner: CVE-2017-7184: xfrm_user: validate XFRM_MSG_NEWAE incoming ESN size harder
Kees Cook has pointed out that xfrm_replay_state_esn_len() is subject to
wrapping issues.  To ensure we are correctly ensuring that the two ESN
structures are the same size compare both the overall size as reported
by xfrm_replay_state_esn_len() and the internal length are the same.

CVE-2017-7184
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Steffen Klassert <steffen.klassert@secunet.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Change-Id: I4fcf9507a2f652b62918261b681c6e921852f55f
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/302386
Reviewed-by: joe_tsai <joe_tsai@asus.com>
Tested-by: joe_tsai <joe_tsai@asus.com>