arm64: dts: htc: avoid /delete-node/ in audio-muskie.dtsi
msm_cdc_pinctrl@59, msm_cdc_pinctrl@60 are defined in
qcom/msm8998-audio.dtsi as devices whose compatible
string is "qcom,msm-cdc-pinctrl". The driver
drivers/mfd/msm-cdc-pinctrl.c is a platform driver.
Therefore, setting the "status" property of both
nodes to "disabled" prevents the device to be probed,
which achieves the same goal as /delete-node/.

Same appiles to &slim_aud/tasha_codec/swr_master and
&slim_aud/tavil_codec/swr_master, with a driver that
checks the availability of the node.

Change-Id: I70873336b6905c7412cbaa8a1829a58871fa4898
Signed-off-by: Yueyao Zhu <yueyao@google.com>