vhost/scsi: potential memory corruption
Fixes CVE-2015-4036

This code in tcm_vhost_make_tpg() is confusing because we limit "tpgt"
to UINT_MAX but the data type of "tpg->tport_tpgt" and that is a u16.

I looked at the context and it turns out that in
vhost_scsi_set_endpoint(), "tpg->tport_tpgt" is used as an offset into
the vs_tpg[] array which has VHOST_SCSI_MAX_TARGET (256) elements so
anything higher than 255 then it is invalid.  I have made that the limit
now.

In tcm_vhost_send_evt() we mask away values higher than 255, but now
that the limit has changed, we don't need the mask.

Change-Id: Ic25b5ae4b76286c332798e53d0cffcd7ec772e30
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>
Signed-off-by: gwx419604 <gwx419604@notesmail.huawei.com>