UPSTREAM: tracing: Add support for critical section event tracing
Critical section trace events will be used for tracing the start and
end of a critical section which can be used by a trace viewer like
systrace to graphically view the start and end of a critical section and
correlate them with latencies and scheduling issues.

Reason for starting critical section:
Either preemption or interrupts were turned off.
Reason for ending critical section:
Both preemption and interrupts are now turned back on.

Bug: 37118023

Change-Id: I28284bc52b63b10d91c85c55f8fc191ed8cb360e
Signed-off-by: Joel Fernandes <joelaf@google.com>