drv2624: make interrupt gpio optional
The TRIG-INIT pin on drv2624 is optional to have in case if the host
choose to be notified on the status change described in the INTZ register.

This change also makes change to use devm_ variant of the gpio_request()
so that we don't need to worry about freeing them in the remove.

Bug: 36707282

Change-Id: I911109bb8fe4a74037801205c22ba2126ecc769c
Signed-off-by: David Lin <dtwlin@google.com>