input: touchscreen: synaptics: reflash whenever config version is different
Currently a full reflash is only triggered when the firmware version is
different or the config is greater. Since walleye config versions are
always greater than those of muskie, a muskie device accidentally flashed
with a walleye build may update to walleye firmware/config and will not
automatically recover when reflashed with a muskie build.

Change-Id: I67c8e3e753fb04e9b17d95710ec46980f2afb348
Signed-off-by: Steve Pfetsch <spfetsch@google.com>