smb-lib: continue overheat mitigation until cable unplugged
When the overheat mitigation is running, and the temperature
has dropped below the threshold, probe the cc status to see
if the cable is unplugged. Ends the mitigation only when
the cablue is unplugged.

The check is done by briefly enabling the Type-C block as
Type-C power role sink, wait for a delay to let the Type-C
block to settle and check the CC attachment status. Disable
the Type-C block at the end.

Bug: 35398639
Change-Id: I3e5f3d247441597ddad7b5fbf78a5af870a1d933
Signed-off-by: yihsiang_peng <yihsiang_peng@htc.com>
Signed-off-by: Dojip Kim <dojip.kim@lge.com>
Signed-off-by: Yueyao Zhu <yueyao@google.com>