smb-lib: hold wakelock when port overheat monitor is running
From the start of the overheat monitor (charging status change),
until the overheat monitor stops (i.e. none of the three conditions
are met, 1) the device is still being charged OR 2) overheat is
currently happending OR 3) an error has occurred), a wake lock is
held to prevent the system from suspend. Otherwise, the overheat
monitor work may not be run in time to perform mitigation.

TEST:
- charge the device (workstation USB)
- - monitor start running (smblib log: "Port overheat work running")
- echo 800 > /sys/class/power_supply/usb/port_temp
- - heat mitigation starts
- unplug cable
- - monitor still runs, system does not suspend
- echo 250 > /sys/class/power_supply/usb/port_temp
- - heat mitigation ends (log: "Port overheat mitigation ends.")
- - system goes to suspend

Bug: 35398639
Change-Id: Idf1e60ec3ab385a4a2203f2ba3294c0820cd0bf1
Signed-off-by: yihsiang_peng <yihsiang_peng@htc.com>
Signed-off-by: Dojip Kim <dojip.kim@lge.com>
Signed-off-by: Yueyao Zhu <yueyao@google.com>