Fix Kbuild to enable WLAN debugfs support
Currently, the CONFIG_WLAN_POWER_DEBUGFS is not enabled for
statically loaded driver. This results in missing subsystem power
stats.
The reason for that is the enabling of this macro is conditionally
done under KERNEL_BUILD=0. This commit enables the macro to be also
conditionally done under all values of KERNEL_BUILD

Bug: 29339696
Test: manual
Change-Id: Ifef7fb7a407a2817134622ef0e760ccbe6343038
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>