USB: dwc3: Switch to TPS only for audio device
This CL causes the switch from PMIC to TPS only when the attached
device has atleast one audio interface.

The switch from TPS->PMIC seems to be causing voltage glitches on
VBUS making the remote device to think that the host is disconneting.
This causes accessory mode to go on reset cycles and causing it to
eventually fail at times.

Bug: 37072175
Change-Id: I542339b632121a45917176f2614f91bf07251805
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>