FROMLIST: binder: protect against stale pointers in print_binder_transaction
(from https://patchwork.kernel.org/patch/9817761/)

When printing transactions there were several race conditions
that could cause a stale pointer to be deferenced. Fixed by
reading the pointer once and using it if valid (which is
safe). The transaction buffer also needed protection via proc
lock, so it is only printed if we are holding the correct lock.

Bug: 36650912
Change-Id: Idc1f18ba59a92a8f2dcc0c012a5a72806d6e84b5
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>