FROMLIST: msm: ipa3: fix channel stop retry logic
Stopping a PROD channel might result in a timeout because the
channel is not empty, or IPA is busy. Whether or not to retry
a stop operation is determined by channel owner.
This change removes the common retry logic of retrying to stop
a PROD channel and adds a retry only for the scenario it is
actually needed.

Change-Id: I7ac1e81f7f99de2b0c3162aa5aaea2102a450838
CRs-Fixed: 2037955
Acked-by: Ady Abraham <adya@qti.qualcomm.com>
Signed-off-by: Skylar Chang <chiaweic@codeaurora.org>
(am from
https://source.codeaurora.org/quic/la//kernel/msm-4.4/commit/?id=a3488e9398c42d1a9046c770c80ef3f81b527404)
Signed-off-by: Yueyao Zhu <yueyao@google.com>
Bug: 63832069