bpf: map_get_next_key to return first key on NULL
commit 8fe45924387be6b5c1be59a7eb330790c61d5d10 upstream.

When iterating through a map, we need to find a key that does not exist
in the map so map_get_next_key will give us the first key of the map.
This often requires a lot of guessing in production systems.

This patch makes map_get_next_key return the first key when the key
pointer in the parameter is NULL.

Change-Id: I2c9a6c017fc315e5228df99970b7ac91b80def99
Signed-off-by: Teng Qin <qinteng@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Chenbo Feng <fengc@google.com>
Cc: Lorenzo Colitti <lorenzo@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Git-commit: fcbc8d0e7dbef92a6b611a6d3d1ed8ea228464a0
Git-repo: https://android.googlesource.com/kernel/common
[schikk@codeaurora.org: commit is taken from msm-4.9 and resolved
trivial merge conflicts]
Signed-off-by: Shadab Naseem <snaseem@codeaurora.org>
Signed-off-by: Swetha Chikkaboraiah <schikk@codeaurora.org>