2017-04-partner: CVE-2017-6424: A-32086742: wlan: Fix buffer overflow in WLANSAP_Set_WPARSNIes()
qcacld-2.0 to prima propagation

Currently In WLANSAP_Set_WPARSNIes() the parameter WPARSNIEsLen
is user-controllable and never validates which uses as the length
for a memory copy. This enables user-space applications to corrupt
heap memory and potentially crash the kernel.

Fix is to validate the WPARSNIes length to its max before use as the
length for a memory copy.

Change-Id: I6edd1988809908f96a4e06d6a8a8678412da12f6
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/295407
Reviewed-by: joe_tsai <joe_tsai@asus.com>
Tested-by: joe_tsai <joe_tsai@asus.com>