power: qpnp-smb2: remove workaround for USBIN supplied VCONN
VCONN is supplied by USBIN on MTP8998 which causes removal issues when
PMI8998 is in source mode and VCONN is enabled. The workaround is
to ensure VCONN is disabled when OTG is disabled which has implications
on PD power role swap when PMI8998 is sourcing VCONN.

The reference schematic has been updated to address this issue. Remove
the workaround.

Change-Id: Ibe2309f3618d840340bd083cca0df8a8644dd9c8
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>