arm64/dts: taimen: drv2624: update tuning parameters
This sets on-resonance frequency for both closed-loop and open-loop mode
to 155 Hz as well as setting the waveform playback interval to 1 ms for
better waveform playback precision. Signal shape is also changed to
square wave to provide a stronger feedback.

The default calibration data (bemf, gain and comp) is also updated here
for devices that did not go through calibration in the factory.

Bug: 37952639

Change-Id: Ie04ebc9d7943cdb14cdbf0daf251b5ad0980e1ba
Signed-off-by: David Lin <dtwlin@google.com>