misc: mnh: MIPI Rx-DPHY bug fix
This patch addresses a bug in the MIPI Rx-DPY IP
which can result in clock lane high speed transition misses
and therefore dropped frames.
It implements a temporary increase of the supply voltage each
time the link is established. DPHY register polling is used to
determine when the DPHY has entered HS. When HS is reached the
supply voltage is reduced back to nominal.

Bug: 37987555
Change-Id: I47e178b0815c89f5dcde0e342a54d3ad72258e33
Signed-off-by: Arnd Geis <arndg@google.com>