defconfig: msmcortex: Disable CONFIG_FORCE_ALLOC_FROM_DMA_ZONE
This flag is primarily being used for WLAN to support WLAN hardware
which cannot handle "> 32 bit" physical addresses for DMA.

Disable this flag for the perf variant of msm8998. This flag forces DMA
addresses to be in 32 bit region. Using this flag with ihelium WLAN
solution (which can handle 64 bit physical DMA addresses) is causing a
reduction in peak KPIs for TCP RX.

Since kernel 4.4 is being used only by ihelium and not being used
for older versions of WLAN hardware (e.g. ROME), we can safely disable
this flag.

CRs-Fixed: 2044326
Change-Id: If2e7d4bab15f6211e91b1095a8ac6c2bf7e9000e
Signed-off-by: Mohit Khanna <mkhannaqca@codeaurora.org>