common: filesystem: limit filesystem stacking depth
Add a simple read-only counter to super_block that indicates how deep this
is in the stack of filesystems.  Previously ecryptfs was the only stackable
filesystem and it explicitly disallowed multiple layers of itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

This patch doesn't contain fs/overlayfs/super.c, the reason is this branch
never uses overlayfs.
CVE: CVE-2014-9922
Bug: 32761463

Change-Id: I6d6338b4982346c9e21bf17304afdc3c9b9b1c9b
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/296998
Reviewed-by: tc_tseng <tc_tseng@asus.com>
Tested-by: tc_tseng <tc_tseng@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>