wlan: Add support to honour multiple scan plans
Add Support to configure multiple 'scan plans' for scheduled scan.
Each 'scan plan' defines the number of scan cycles and the interval
between scans. The scan plans are executed in the order they were
configured. The last scan plan will always run infinitely and thus
defines only the interval between scans
The maximum number of scan plans supported by the device and the
maximum number of iterations in a single scan plan are advertised
to userspace so it can configure the scan plans appropriately.

Change-Id: Ib82594a9e9be88ac7c36e1804e84b21972f1404b
CRs-Fixed: 1099024