Revert "drivercore: Avoid adding devices without pm_ops to dpm list"
Commit d36bb30811b ("drivercore: Avoid adding devices without pm_ops to
dpm list") keeps the devices that do not have pm callback out of the PM
core's active device list in order to optimize the suspend / resume time
by skipping those devices.

This does not work as drivers with pm callback may be bounded to the
devices at a later time. This is currently causing those drivers with pm
callbacks support not being called in the current tree.

Plus this optimization does not help the performance much as the pm core
is already using power.no_pm_callbacks to allow devices without pm
callbacks to go direct_complete.

This reverts commit d36bb30811b9ea1467f9b10a5d701240230147b9.

Bug: 38248799
Change-Id: I015f7e0505dd16eb931f5844cf832d0bc2bebaa7
Reported-by: Carter Hsu <carter_hsu@htc.com>
Signed-off-by: David Lin <dtwlin@google.com>