qcacld-3.0: Stale scan entries updated to NL
Stale scan entries are updated to NL because
scan results are not aged out in PNO complete
handler.
Age out scan results after receiving PNO
complete notification to CSR and not update
scan results as part of sched scan callback
as bss details are updated to NL as part of
csr_scan_process_single_bssdescr.

qcacld-3.0: Do not flush scan results for sched scan

Scan results are flushed as part of sched scan start
which could result in subsequent connect to fail as
bss is removed from scan list.
Fix is to not flush scan results as part of sched
scan start.

Bug: 38028945
Bug: 37758502
Change-Id: I2ed6b7b8c6e27daf921f5804eea6a12f72493044
Signed-off-by: Subhani Shaik <subhanis@codeaurora.org>