asoc: msm: add msm8998 based wahoo machine driver
Commit cda1cb26c80a19e9 ("ASoC: msm: Add loadable module support for
machine driver") split the msm8998 machine driver from the platform
driver. We could simply add the msm8998 machine driver config back,
however, this is not desired as msm8998 machine driver is intended to be
used for msm8998 MTP reference board which uses wsa881x driver that's
not used on the Wahoo board.

This patch separates the Wahoo ASoC machine driver by cloning the
msm8998 MTP machine driver and remove all the reference to the wsa881x
codec.

Change-Id: I504e4ac549dcf446bb53d250aca76ca8b68e5306
Signed-off-by: David Lin <dtwlin@google.com>