arm64/configs: marlin: Enable CONFIG_DEBUG_OBJECTS_TIMERS
We're seeing crashes due to an hrtimer red-black tree getting
corrupted and a hrtimer node's left or right pointer points to itself.
The crash manifests as a CPU infinite looping with interrupts disabled
while iterating over the corrupted hrtimer red-black tree in
enqueue_hrtimer() or remove_hrtimer().

This is a debugging patch to detect certain types of hrtimer corruption due to:
1. Improper use of the hrtimer API like starting an uninitialized timer.
2. Freeing memory used to store an hrtimer while the timer is still active.

Bug: 38121882
Change-Id: I65b51ec35effe103ddd4b8bac4f259cedce2a00c
Signed-off-by: Siqi Lin <siqilin@google.com>