rbtree: Detect left/right pointer assignments that would cause loops
Add checks before all left/right pointer assignments to verify that
the pointer value being assigned isn't the same node as the left/right
pointer. This is a BUG_ON() because we want to get a crash dump as
close to the source of the corruption as possible. Once the red-black
tree is corrupted, the kernel will eventually crash anyway and by
then the root cause will be gone.

Bug: 38121882
Change-Id: Ifa59c08d5bd2218e667e4c08162fffbaf388628a
Signed-off-by: Siqi Lin <siqilin@google.com>