FROMLIST: binder: use atomic for transaction_log index
(from https://patchwork.kernel.org/patch/9817807/)

The log->next index for the transaction log was
not protected when incremented. This led to a
case where log->next++ resulted in an index
larger than ARRAY_SIZE(log->entry) and eventually
a bad access to memory.

Fixed by making the log index an atomic64 and
converting to an array by using "% ARRAY_SIZE(log->entry)"

Also added "complete" field to the log entry which is
written last to tell the print code whether the
entry is complete

Bug: 62038227
Change-Id: I3c4223fd4cb6496061c75b5d817b643995151703
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>