misc: easelcomm: Add timeout for flush
When registering a service, client side will request remote to
flush messages.  This will be performed by the remote in kernel driver
and should always do flush.  Client waits for the reply from remote,
blocking the EASELCOMM_IOC_REGISTER ioctl call.

In case the remote kernel is by accident in a bad state, or messages
are ignored by remote, but power is still on, we add an arbitrary
5 sec timeout for service register.

Also, complete flush_done immediately if remote is shutdown while
local is waiting for flush_done completion.

Bug: 38507689

Change-Id: Idb0b8530edc7548f2b2879b05f90035d0e92c23d
Signed-off-by: Cheng Gu <gucheng@google.com>