usb: fusb302: Change data role to host when port state is disconnected
Having data role set to device when port in disconnected causes the
port to enumerate even before TCPM settles in ATTACHED_SNK state.
This causes the adb notification to race against the TYPEC notification.
Leave the port in HOST role when the state is disconnected, so that
data path is not enabled unless the port settles in ATTACHED_SRC state.

Bug: 37094615
Change-Id: I36e2dfe63a8489a7e95a3e2585a7f5490eef9a25
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>