PD: policy_engine: Add timeout for mode change through sysfs
Currently USB driver tries to set power role as none and waits for
disconnect notification to set new mode. But there is a chance that it
might fail and stuck in infinite loop if other device is not dual-role
capable. Fix it by having timeout and bail out in case timeout happens.

Change-Id: I569356ee445c53950ed8fdd3f9994d4ae0e86131
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>