usb: typec: allow tcpm logs to overflow
TCPM logging uses a in memory ring buffer. When the ring buffer
is full, currently it writes a string 'overflow' to the last slot.
This can be useful during debug to catch for early errors before the
state machine goes into a loop and floods the log.

However, in a bugreport from users, most logs will be overflown in
this setting. Therefore, allow the ring buffer to overflow
(overriding the old logs) to capture the latest logs.

Change-Id: I40e2fa6ab955f189e5f575624beadadbbae86a84
Signed-off-by: Yueyao Zhu <yueyao@google.com>