ARM: dtsi: msm: reserve memory for LK's stack on msm8996
The memory node LK's stack is running on should be reserved
in system booting stage. Otherwise, the processes like early
domain service on LK will be impacted. And in order not to
impact common dtsi file for mobile and other platforms,
only put the change in this auto-specific dtsi file.

Change-Id: Ica8d2a98e00289990789eedf6f4e3d88b03acfcc
Signed-off-by: Guchun Chen <guchunc@codeaurora.org>