wlan: Increment scan failure counter only on connection  in progress
Presenly, along with connection in progress scenario, scan failure
counter is also incremented on BTcoex operation is in progress,
leading counter to reach more than 35 not due to connection in
progress state but due to BTCoex mode in progress.

Hence increment scan failure counter only when connection is in
progress.

Change-Id: I4f47909d05f942bef825776844fa136ba7ade661
CRs-Fixed: 1066610