drm/msm : fix hdmi controller register programming
Currently, there is an incorrect programming of the
hdmi controller register where default value of the
datapath mode bit is being toggled along with the
scrambler enable/disable.

This bit should be untouched as per the hardware
programming sequence and kept at the default state.

Fix the register programming to avoid toggling this
bit.

Change-Id: I2e8f74c0abaddd27b5d8a2136afc4e1b82f96e7d
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>