security: ion: Fix use after free during ION_IOC_ALLOC
If a user happens to call ION_IOC_FREE during an
ION_IOC_ALLOC on the just allocated id, and the
copy_to_user fails, the cleanup code will attempt
to free an already freed handle.

This adds a wrapper for ion_alloc that adds an
ion_handle_get to avoid this.

Bug: 31568617
Change-Id: Ib7eeb3ef067d8b9062f275e5d394a7667a71dd09
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/303078
Reviewed-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Tested-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>