misc: easelcomm: Use vmalloc for local sg
There are cases that ION buffer scatterlist has thousands of entries
while kernel is out of order 5, making it failed to allocate memory
for dma_xfer.sg_local.

Since the input of numbers of entries is not predictable, and the
content of sg_local is only accessed by CPU, not the DMA hardware,
it does not have to be physically continuous memory.

This patch converts dma_xfer.sg_local to vmalloc from kmalloc.

Bug: 62587167
Change-Id: I3099557a7876914ccf9fee00d68d3ed060057894
Signed-off-by: Cheng Gu <gucheng@google.com>