arm64/configs: taimen: reduce zc-dec time to 100uS
This patch fixes an issue with haptics noise. A longer ZC_DET_TIME will
cause the drive time to be slightly shorted.  The way the algorithm
works is that it subtracts the sum of IDISS, BLANKING, and ZC times from
the drive time. So changing the ZC_DET_TIME will give a slightly shorter
drive time and longer guard ring time. This may case the rated voltage
(Vrms) to change based on the algorithm timings.

Bug: 63270286
Test: long haptics on notification
Change-Id: I7e2c6930c52f885300bdf03478d8881d90db5e6d
Signed-off-by: David Lin <dtwlin@google.com>