ASoC: da7219: Enable MCLK only when it is needed.
The MCLK is only required to enable audio playback on the DAI
link. Therefore, only enable it when the DAI link is enabled.
Previously the MCLK was enabled when an associated audio card
was enabled for any audio playback. For example, the card is
enabled to playback to speakers via another DAI link, and the
DAPM BIAS is set ON for all devices on that card.

Bug: 63823804
Test: Playback functions as expected. MCLK is enabled for
      playback or capture over the headset jack only. Headset
      jack detection and button presses function normally.

Change-Id: If3852f3588a8597522acac42de91fe32ab27a962