dccp: fix freeing skb too early for IPV6_RECVPKTINFO
CVE-2017-6074

In the current DCCP implementation an skb for a DCCP_PKT_REQUEST packet
is forcibly freed via __kfree_skb in dccp_rcv_state_process if
dccp_v6_conn_request successfully returns.

However, if IPV6_RECVPKTINFO is set on a socket, the address of the skb
is saved to ireq->pktopts and the ref count for skb is incremented in
dccp_v6_conn_request, so skb is still in use. Nevertheless, it gets freed
in dccp_rcv_state_process.

Fix by calling consume_skb instead of doing goto discard and therefore
calling __kfree_skb.

Change-Id: Ie1b4f85e4753a941eb00ac2b01185fdf23ef0e56
Signed-off-by: ywx486968 <ywx486968@notesmail.huawei.com>