ipv4: keep skb->dst around in presence of IP options
CVE-2017-5970

Andrey Konovalov got crashes in __ip_options_echo() when a NULL skb->dst
is accessed.
ipv4_pktinfo_prepare() should not drop the dst if (evil) IP options
are present.
We could refine the test to the presence of ts_needtime or srr,
but IP options are not often used, so let's be conservative.
Change-Id: Ie58b4af0e1b13d01cc8b757108072455672e3cc5