hwmon: max31760: Restore manual PWM after reset/resume.
The direct duty-cycle control reigster (PWMR) can set the PWM
when the device is in manual control mode. However, when the
device is reset this value returns to 255, and when the device
resumes from standby, this value returns to 0.

Use one of the user-programmable registers to store the PWMR set
by the user, and restore that value at probe time, and after
resume from sleep.

Bug: 63999060
Test: Fan controller sleeps properly.
      In manual mode, PWM value returns to the value set after
      resume, and after system restart.
Change-Id: Ibf13dee1329ea78c4ac0c8bf28f08715efc61959