AndroidKernel.mk: Remove excessive $(hide)s
The "@" symbol is used within Makefile rules to silence the output
of the command as make is processing it.  However, this is not
useful in a compound command such as "@ if [ xxx ]; @ b ; fi"
since make neither sees nor removes the second @.  The bash
interpreter doesn't understand this "@" and will throw an error.

Bug: None
Test: m -j32
Change-Id: I1e2d77304b40e102a0ddcc50fd4a7b0be6fb1284