usb: type-c: tcpm: Check for Rp for tPDDebounce
According the spec, the following is the conditions for exiting Try.SNK
state:
"The port shall wait for tDRPTry and only then begin monitoring the CC1 and
CC2 pins for the SNK.Rp state. The port shall then transition to
Attached.SNK when the SNK.Rp state is detected on exactly one of the CC1
or CC2 pins for at least tPDDebounce and V BUS is detected. Alternatively,
the port shall transition to TryWait.SRC if SNK.Rp state is not detected
for tPDDebounce."

Bug: 64261058
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Change-Id: I2a12a833d0b2d9718a49f7bee53ceca8382ea2ef