FROMLIST: pstore: Make default pstorefs root dir perms 0750
Currently only DMESG and CONSOLE record types are protected, and it isn't
obvious that they are using a capability check. Instead switch to explicit
root directory mode of 0750 to keep files private by default.  This will
allow the removal of the capability check, which was non-obvious and
forces a process to have possibly too much privilege when simple post-boot
chgrp for readers would be possible without it.

Bug: 117834054
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Siqi Lin <siqilin@google.com>
(am from https://patchwork.kernel.org/patch/9894577/)
Signed-off-by: Jonglin Lee <jonglin@google.com>

Change-Id: Ie7ceedc152bcaf17d8eaece2cfffbae1b432f45f