Trust Polaris release certificate for verified boot.
If this cert is not in the system key ring and the system partition is
signed with a release key, android-verity will refuse to mount "/".

Appending the cert in the .pem file allows android-verity to trust
either certificate. Although the kernel can trust either certificate
(which allows the same kernel binary to be used for both eng and user
builds) the bootloader will pin down which certificate can be used on
the kernel command line.

Bug: 64577024
Test: none
Change-Id: Ifa5f15c0fb2fec7399e72dcd7f57894950bbeeb7