drm/msm: Map buffers on demand on the submit path
Map and pin buffers on demand on the submission path.
This ensures that we only map buffers whose iova are
actually needed for submission as opposed to all
buffers in the buffer list. For instance, the command
buffers, and the reloc buffers for processing relocs.
Also remove unused member valid from the struct
msm_gem_submit.

Change-Id: I644f44f202552d14762ffe1d1761b98ed5961020
Signed-off-by: Sushmita Susheelendra <ssusheel@codeaurora.org>