drv2625: verify GO bit after writing
After writing to the GO bit to turn off the vibrator, read back
the value until it is 0, to allow auto-braking to finish before
the vibrator can be turned on again.

Also configure the device's PLAYBACK_INTERVAL setting to 1ms to
allow for crisper short vibration bursts and shortening the
auto-break delay to 10ms.

This change should fix the 'vibrator stuck' issues reported.

Bug: 62450511
Bug: 64424253
Bug: 64033422

Change-Id: Ief16014973d5b58856b84d3be362f4b7a439ab41
Signed-off-by: Matthias Hausner <hausner@google.com>