arm: dts: angelfish: Recharge threshold of 150 mA
On some devices, the prior recharge threshold resulted in the recharge
period being only a few minutes long.  Constant wakeups to disable and
enable wireless power hurt efficiency, heated the device, and put extra
cycles on the battery.  With this change the battery is allowed to drop
to 4250 mV (97%) before recharging is initiated which stretches out the
recharge period.

Bug: 62631993
Change-Id: I6bde53fc014fc231e5862b65a8316cc83a9d3cd1
Signed-off-by: Tim Kryger <tkryger@google.com>