arm64: dts: b1c1: fix irq pull settings for ST21NFC chip
IRQ_OUT is a bi-directional pin, not an output.
As such, it does not drive IRQ_OUT always, and when it does not, it
expects particular value at its input, otherwise ST21NFC does not
work as expected.

We have to replace pull-up with pull-down on IRQ_OUT pin of ST21.

Bug: 64202676
Change-Id: Ic2e643f126eeb76d059e33aa4a8d7f7c413eb24b
Signed-off-by: Alexey Polyudov <apolyudov@google.com>