power: qpnp-fg-gen3: adjust ki coefficient for full soc dynamically
In certain conditions, after the battery SOC reaches 100%, SOC is
not getting updated during discharging. Based on the hardware
recommendation, adjust the Ki coefficient for full SOC so that
SOC can be updated faster during discharging.

While at it, remove prev_charge_status variable which is not
used really.

CRs-Fixed: 2089555
Bug: 64633491
Change-Id: I54b7d86529c21c016ab777053895ccda9c7a2b35
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Ajay Dudani <adudani@google.com>