dts: vega: Revert pinctrl fix for DVT
Vega DVT is not compatible with the pinctrl settings for the
camera GPIOs in the same way that Vega DVB is.  This CL creates
a separate DTS file for the DVT and EVB revisions of Vega.  In
doing so, we use the proper board id GPIOs which were not in
the DTS files before.  This CL also reverts the pinctrl change
for the DVT.  As a future, but much lower priority exercise, we
can delve further into re-instating the pinctrl for DVT after
root-causing the problem.

Bug: 65248436
Test: On each device, run:
      adb shell "revision; dumpsys media.camera | grep 'of camera'"
    Vega DVT should say "vega-dvt", Vega EVB should say "vega-dvb"
    and both should say "Number of camera devices: 1"

Change-Id: I15b69196f174cc761ae5b06b55068d5064fc566b