power: smb: Add flag to disable 12V chargers.
Add max_9v_adapter flag to the smb-lib charger structure. When
true, the code will allow a maximum of 9V adapters. Defaults to
false.

Add a 'qcom,max-9v-adapter' boolean property to the qpnp-smb2
driver which sets the max_9v_adapter flag to true when present.

For SDM660 devices, set max_9v_adapter to true to keep the
existing functionality.

Bug: 64703341
Test: Kernel compiles. Flag is available. When set in DTS, 12V
      adapaters are disallowed/downgraded to 9V.
Signed-off-by: John Muir <muirj@google.com>
Change-Id: I779e647f862c54b089f28a6344b7542ee6a3f13a