sched: add prev task to long-runnable log
The logging that was added for detecting when
a task is runnable for a long time is useful
but we could use some more context on what was
running instead. A full dump might be too costly;
adding the previous task to the log message
so we have some idea of what else is running.

picked from android-msm-wahoo-4.4

Bug: 72123250
Change-Id: I7b3c6941caa7fb9cacde216e8a3b1c6c251402f1
Signed-off-by: John Dias <joaodias@google.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>