qcacld-3.0: Fix memory leak for txrx_fw_stats cmd
qcacld-2.0 to qcacld-3.0 propagation

The txrx_fw_stats cmd will allocate a req object before sending the cmd
to the firmware, this memory is only freed when get response from firmware.
The memory leak will appear if the firmware doesn't response in time before
the driver unloading.

This fix will add a pending queue to trace this req object in the pdev.
when pdev is detaching, it will clean up this queue to avoid memory leak.

Bug: 72957257
Change-Id: I35f6216d35befbab978bba161252b305488bd34c
CRs-Fixed: 2116408
Signed-off-by: Ecco Park <eccopark@google.com>