qcacld: Disable wait for driver startup completion.
The convoluted WiFi driver startup will block the entire system
start while the driver waits to communicate with the modem DSP,
which at this point hasn't been started by the cnss-daemon. This
leads to a failure to start the WiFi driver.

While waiting for a better fix from QCOM, we disable the wait
here. This allows the system startup to continue, the modem DSP
is loaded, and WiFi starts.

Bug: 66812758
Test: WiFi starts.
Change-Id: I94919721771e18442f6f651f535c7b933694fb80