Revert "arm64: move ELF_ET_DYN_BASE to 4GB / 4MB"
This ARM mmap change breaks AddressSanitizer:

Shadow memory range interleaves with an existing memory mapping. ASan cannot proceed correctly. ABORTING.

Revert it until ASAN runtime library is updated to handle it.

It can wait as it is rearranging the address space to make certain
exploits harder, not fixing a bug.

Bug: 67425063

This reverts commit b1a269f4564d48ade43acde11a55dc2a06b63476.

Change-Id: Ib6916edf9be0cb8c2bd4073625fd845c0ea08905