ARM: dt: b1c1: B1 P1.1 POLED_TE on GPIO12
Move POLED_TE to GPIO12 on P1.1 B1.

This means some reorg of NFC pins as well, so set up the nfc pinctrl
for P1.1, disable NFC on P1.1, and move the P1 and devboard NFC pins
to match in their respective device trees.

qcom,platform-te-gpio isn't used.  Instead, use the
qcom,platform-vsync-source to define the configuration (sdm845
gpio10 is 0, gpio11 is 1, gpio12 is 2.)

Bug: 68239816
Change-Id: I9d1e34b40789bde4e22e4bc855affd3b10ce3424
Signed-off-by: Chris Fries <cfries@google.com>