input: touchscreen: stm: read touch events in one shot
The incremental i2c reads from the touch controller and event processing
by the CPU was heavily interlaced. Worse, the work was often preempted,
adding substantial latency. Since it sometimes took greater than a
single touch report period to process a frame of events, the touch
controller would start dropping events, particularly in multitouch use
cases.

This approach reads as much of the FIFO as it can in a single read and
then allows the CPU to process all touch events in one shot.

Change-Id: I370b346676b1e5f124e1d5f65dc10edaf094598c
Signed-off-by: Steve Pfetsch <spfetsch@google.com>