Change touch panel dimension configuration to 320x320.
The previous configuration was 320x290, which is the size of the display
since the display does not include the chin of the device. However, the
touch panel actually does extend over the chin and the result was that
the scaling was incorrect so that touches at the bottom of the screen
get registered as lower than they actually are.

Bug: 23869671
Test: Touching the screen with show pointer location on to see that the
locations are accurate

Change-Id: Ie724a4efa69812f19ed265cb81d8b28f1c3f111f